Address Sonar issues
